separate the plotting class to separate module
